Separate Docker Compose services for build, test, run
Refactor Docker Compose setup to include distinct services for building, testing,
and running the application.

Use separate Dockerfiles and configurations for each phase to isolate concerns and
improve workflow.

Quiet down the maven logs and set log levels to WARN so there is less noise in the
logs.

Update README file with information on running the project.